+++ src/main/java/com/akshit/treading/service/WalletService.java
@@ -8,8 +8,8 @@ import com.akshit.treading.modal.Wallet;
 public interface WalletService {
 
     Wallet getUserWallet(User user);
-    Wallet addBalance(Wallet wallet, Long amount);
+    Wallet addBalance(Wallet wallet, double amount);
     Wallet findWalletById(Long id) throws Exception;
-    Wallet transferFunds(User sender,Wallet receiver,Long amount) throws Exception;
+    Wallet transferFunds(User sender,Wallet receiver,double amount) throws Exception;
     Wallet payOrderPayment(Order order, User user) throws Exception;
 }
